Moved tracking helpers into a subfolder, pulled intcmp and refpvid into a guTracking module.
