--- assets/helpers/intCmp.js
@@ -1,10 +1,0 @@
-// @flow
-
-// ----- Reducer ----- //
-
-// Since nothing change the INTCMP, this reducer does not handle any action.
-const intCmpReducer = (state: ?string = null): ?string => state;
-
-export {
-  intCmpReducer,
-};
