--- assets/helpers/refpvid.js
@@ -1,9 +1,0 @@
-// @flow
-
-// ----- Reducer ----- //
-
-const refpvidReducer = (state: ?string = null): ?string => state;
-
-export {
-  refpvidReducer,
-};
