+++ assets/helpers/tracking/guTracking.js
@@ -1,0 +1,17 @@
+// @flow
+
+// ----- Reducers ----- //
+
+// Since nothing changes the INTCMP, this reducer does not handle any actions.
+const intCmpReducer = (state: ?string = null): ?string => state;
+
+const refpvidReducer = (state: ?string = null): ?string => state;
+
+
+// ----- Exports ----- //
+
+export {
+  intCmpReducer,
+  refpvidReducer,
+};
+
