+++ assets/pages/bundles-landing/reducers/reducers.js
@@ -5,7 +5,7 @@
 import { combineReducers } from 'redux';
 
 import { abTestReducer as abTests } from 'helpers/abtest';
-import { intCmpReducer as intCmp } from 'helpers/intCmp';
+import { intCmpReducer as intCmp } from 'helpers/tracking/guTracking';
 import type { Contrib, ContribError, Amounts } from 'helpers/contributions';
 
 import { parse as parseContribution } from 'helpers/contributions';
